moves game ended and open views into separate components
